requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
pydantic-settings
python-dotenv

httpx
async-lru